refactor : statisticsService 분리
