+++ src/index.js
@@ -1,15 +1,10 @@
 import { raceInit, raceManager } from "./service/raceService.js";
 import { OutputView } from "./view/OutputView.js";
+import { SystemMessage } from "./settings/SystemMessage.js";
+import { getWinner } from "./service/statisticsService.js";
 
 const { cars, round } = await raceInit();
 raceManager.startRace(cars, round);
 
-const maxPosition = Math.max(...cars.map((car) => car.position));
-
-const winner = cars
-  .filter((car) => car.position === maxPosition)
-  .map((car) => car.name)
-  .join(", ");
-//
-
-OutputView.printMessage(`최종 우승자: ${winner}`);
+const winner = getWinner(cars);
+OutputView.printMessage(SystemMessage.WINNER_MESSAGE(winner));
