+++ src/settings/SystemMessage.js
@@ -1,6 +1,6 @@
 export const SystemMessage = {
-  CAR_NAME_MESSAGE:
-    "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n",
-  ROUND_MESSAGE: "시도할 횟수는 몇 회인가요?\n",
-  RESULT_MESSAGE: "\n실행 결과",
+  CAR_NAME_MESSAGE: `경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n`,
+  ROUND_MESSAGE: `시도할 횟수는 몇 회인가요?\n`,
+  RESULT_MESSAGE: `\n실행 결과`,
+  WINNER_MESSAGE: (winner) => `최종 우승자: ${winner}`,
 };
